build: eliminate esbuild-plugin-copy and esbuild-common.js dependency
esbuild-plugin-copy:
That has a number of npm dependencies, not packaged in distros, and
doing this with `fs.copyFileSync()` is just as easy.

esbuild-common.js:
The only significant thing in that custom plugin was the scss hacking.
But that is now obsolete. We can use the upstream esbuild-sass-plugin
directly.

See https://github.com/cockpit-project/cockpit/commit/3f331895b9fece4b